Proyecto/Sprint2/scrapeo/scrapeo_2025/requirements.txt
# Dependencias para el scraper de Beca 18
# Instalar con: pip install -r requirements.txt

requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
lxml>=4.9.0
openpyxl>=3.0.0
# Integración con Power BI
msal>=1.26.0
python-dotenv>=1.0.0